Add files via upload
This program uses a spreadsheet (ethercalc.org) in which the members of the group monthly registers their expenses (for food and for other stuff). After everybody has done that the program sends them an email asking to enter also the amount of money they want to pay for this month. After everybody has done that you have different options to do the paying process. Finally everybody gets an email telling the amount the person has still to pay or gets back from the administrative account.